+++ src/App.js
@@ -2,7 +2,7 @@ import { ThemeProvider } from "styled-components";
 import { GlobalStyle } from "./app/GlobalStyle";
 import { theme } from "./app/theme";
 import { MovieList } from "./features/movies/MovieList";
-import SiteHeader from "./common/SiteHeader";
+import SiteHeader from "./features/SiteHeader";
 
 function App() {
   return (
